requirements.txt
python-telegram-bot>=21.0
python-dotenv>=1.0.0
google-genai>=1.0.0
httpx[http2]>=0.27.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-json-logger>=2.0.7

